+++ CHANGELOG.md
@@ -20,6 +20,9 @@ ChangeLog history
 ## Features
 - a new operator called `save-images` to save chunk as a serials of PNG images.
 ## Bug Fixes 
+- queue becomes None
+- pznet was not working.
+- fix Dockerfile for pznet build (not sure whether breaks PyTorch or not)
 
 ## Improved Documentation 
 - add kubernetes documentation to find monitor of operations per second and bandwidth usage.